put extension modules into their own translation units
