+++ cpp/box/export_Box.cc
@@ -7,8 +7,8 @@ namespace nb = nanobind;
 
 namespace freud { namespace box { namespace wrap {
 
-void makeAbsolute(std::shared_ptr<Box> box, nb_array<float, nb::shape<-1, 3>> vecs,
-                  nb_array<float, nb::shape<-1, 3>> out)
+void makeAbsolute(std::shared_ptr<Box> box, nb_array<float> vecs,
+                  nb_array<float> out)
 {
     unsigned int Nvecs = vecs.shape(0);
     vec3<float>* vecs_data = (vec3<float>*) (vecs.data());
@@ -16,8 +16,8 @@ void makeAbsolute(std::shared_ptr<Box> box, nb_array<float, nb::shape<-1, 3>> ve
     box->makeAbsolute(vecs_data, Nvecs, out_data);
 }
 
-void makeFractional(std::shared_ptr<Box> box, nb_array<float, nb::shape<-1, 3>> vecs,
-                    nb_array<float, nb::shape<-1, 3>> out)
+void makeFractional(std::shared_ptr<Box> box, nb_array<float> vecs,
+                    nb_array<float> out)
 {
     unsigned int Nvecs = vecs.shape(0);
     vec3<float>* vecs_data = (vec3<float>*) (vecs.data());
@@ -25,8 +25,8 @@ void makeFractional(std::shared_ptr<Box> box, nb_array<float, nb::shape<-1, 3>> 
     box->makeFractional(vecs_data, Nvecs, out_data);
 }
 
-void getImages(std::shared_ptr<Box> box, nb_array<float, nb::shape<-1, 3>> vecs,
-               nb_array<int, nb::shape<-1, 3>> images)
+void getImages(std::shared_ptr<Box> box, nb_array<float> vecs,
+               nb_array<int> images)
 {
     const unsigned int Nvecs = vecs.shape(0);
     vec3<float>* vecs_data = (vec3<float>*) (vecs.data());
@@ -34,8 +34,8 @@ void getImages(std::shared_ptr<Box> box, nb_array<float, nb::shape<-1, 3>> vecs,
     box->getImages(vecs_data, Nvecs, images_data);
 }
 
-void wrap(std::shared_ptr<Box> box, nb_array<float, nb::shape<-1, 3>> vecs,
-          nb_array<float, nb::shape<-1, 3>> out)
+void wrap(std::shared_ptr<Box> box, nb_array<float> vecs,
+          nb_array<float> out)
 {
     const unsigned int Nvecs = vecs.shape(0);
     vec3<float>* vecs_data = (vec3<float>*) (vecs.data());
