+++ cpp/locality/export_Filter.h
@@ -10,45 +10,27 @@
 
 #include <nanobind/nanobind.h>
 #include <nanobind/ndarray.h>
-namespace nb = nanobind;
 
 namespace freud { namespace locality {
 
-template<typename T, typename shape>
-using nb_array = nanobind::ndarray<T, shape, nanobind::device::cpu, nanobind::c_contig>;
-
 namespace wrap {
 
+template<typename T, typename shape = nanobind::shape<-1, 3>>
+using nb_array = nanobind::ndarray<T, shape, nanobind::device::cpu, nanobind::c_contig>;
+
 void compute(std::shared_ptr<Filter> filter, std::shared_ptr<NeighborQuery> nq,
-             nb_array<float, nb::shape<-1, 3>> query_points, std::shared_ptr<NeighborList> nlist,
-             const QueryArgs& qargs)
-{
-    const auto num_query_points = query_points.shape(0);
-    const auto* query_points_data = (vec3<float>*) query_points.data();
-    filter->compute(nq, query_points_data, num_query_points, nlist, qargs);
-}
+             nb_array<float> query_points, std::shared_ptr<NeighborList> nlist,
+             const QueryArgs& qargs);
 
 }; // namespace wrap
 
 namespace detail {
-void export_Filter(nb::module_& m)
-{
-    nb::class_<Filter>(m, "Filter")
-        .def("compute", &wrap::compute, nb::arg("nq"), nb::arg("query_points"), nb::arg("nlist").none(),
-             nb::arg("qargs"))
-        .def("getFilteredNlist", &Filter::getFilteredNlist)
-        .def("getUnfilteredNlist", &Filter::getUnfilteredNlist);
-}
 
-void export_FilterRAD(nb::module_& m)
-{
-    nb::class_<FilterRAD, Filter>(m, "FilterRAD").def(nb::init<bool, bool>());
-}
+void export_Filter(nanobind::module_& m);
 
-void export_FilterSANN(nb::module_& m)
-{
-    nb::class_<FilterSANN, Filter>(m, "FilterSANN").def(nb::init<bool>());
-}
+void export_FilterRAD(nanobind::module_& m);
+
+void export_FilterSANN(nanobind::module_& m);
 
 }; // namespace detail
